Add internal etcd to simplify dev setup
Now if you do nothing with the docker run,
it will by default start up an etcd process
inside the container, so you don't have
to run another container for local etcd.

You can still force ringleader to do that with
the ETCD_HOST env var, in which case it
will not start the other etcd process.